fixed strokeThickness

+++ src/MessengerWhiteboard/SerializableShapeItem.cs
@@ -15,7 +15,7 @@ namespace MessengerWhiteboard
 
         public Color color { get; set; }
         public int ZIndex { get; set; }
-        public int StrokeThickness { get; set; }
+        public double StrokeThickness { get; set; }
 
         public List<Point> points{ get; set;}
     }
